+++ pytorch_lightning/plugins/training_type/single_tpu.py
@@ -12,23 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
-from typing import Any, Dict, Optional
+from typing import Optional
 
 import pytorch_lightning as pl
 from pytorch_lightning.plugins.io.xla_plugin import XLACheckpointIO
 from pytorch_lightning.plugins.precision import PrecisionPlugin
-from pytorch_lightning.plugins.training_type.single_device import SingleDevicePlugin
+from pytorch_lightning.plugins.training_type.single_device import SingleDeviceStrategy
 from pytorch_lightning.utilities import _TPU_AVAILABLE, find_shared_parameters, set_shared_parameters
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.model_helpers import is_overridden
-from pytorch_lightning.utilities.types import _PATH
 
 if _TPU_AVAILABLE:
     import torch_xla.core.xla_model as xm
 
 
-class SingleTPUPlugin(SingleDevicePlugin):
-    """Plugin for training on a single TPU device."""
+class SingleTPUStrategy(SingleDeviceStrategy):
+    """Strategy for training on a single TPU device."""
 
     def __init__(
         self,
@@ -74,20 +73,11 @@ class SingleTPUPlugin(SingleDevicePlugin):
     def model_to_device(self) -> None:
         self.model.to(self.root_device)
 
-    def save_checkpoint(self, checkpoint: Dict[str, Any], filepath: _PATH) -> None:
-        """Save model/training states as a checkpoint file through state-dump and file-write.
-
-        Args:
-            checkpoint: dict containing model and trainer state
-            filepath: write-target file's path
-        """
-        return self.checkpoint_io.save_checkpoint(checkpoint, filepath)
-
     def teardown(self) -> None:
         # TPU teardown
         os.environ.pop("PT_XLA_DEBUG", None)
 
-    @SingleDevicePlugin.checkpoint_io.setter
+    @SingleDeviceStrategy.checkpoint_io.setter
     def checkpoint_io(self, io: Optional[XLACheckpointIO]) -> None:
         if io is not None and not isinstance(io, XLACheckpointIO):
             raise MisconfigurationException(f"{self.__class__.__name__}.checkpoint_io` must be a `XLACheckpointIO`.")