+++ tests/callbacks/test_tqdm_progress_bar.py
@@ -17,7 +17,7 @@ import sys
 from collections import defaultdict
 from typing import Union
 from unittest import mock
-from unittest.mock import ANY, call, Mock
+from unittest.mock import ANY, call
 
 import pytest
 import torch
@@ -172,27 +172,16 @@ def test_tqdm_progress_bar_progress_refresh(tmpdir, refresh_rate: int):
         val_batches_seen = 0
         test_batches_seen = 0
 
-        def on_train_batch_start(self, trainer, pl_module, batch, batch_idx):
-            super().on_train_batch_start(trainer, pl_module, batch, batch_idx)
-            assert self.train_batch_idx == trainer.fit_loop.batch_idx
-
         def on_train_batch_end(self, trainer, pl_module, outputs, batch, batch_idx):
             super().on_train_batch_end(trainer, pl_module, outputs, batch, batch_idx)
-            assert self.train_batch_idx == trainer.fit_loop.batch_idx + 1
-            if not self.is_disabled and self.train_batch_idx % self.refresh_rate == 0:
-                assert self.main_progress_bar.n == self.train_batch_idx
             self.train_batches_seen += 1
 
         def on_validation_batch_end(self, trainer, pl_module, outputs, batch, batch_idx, dataloader_idx):
             super().on_validation_batch_end(trainer, pl_module, outputs, batch, batch_idx, dataloader_idx)
-            if not self.is_disabled and self.val_batch_idx % self.refresh_rate == 0:
-                assert self.val_progress_bar.n == self.val_batch_idx
             self.val_batches_seen += 1
 
         def on_test_batch_end(self, trainer, pl_module, outputs, batch, batch_idx, dataloader_idx):
             super().on_test_batch_end(trainer, pl_module, outputs, batch, batch_idx, dataloader_idx)
-            if not self.is_disabled and self.test_batch_idx % self.refresh_rate == 0:
-                assert self.test_progress_bar.n == self.test_batch_idx
             self.test_batches_seen += 1
 
     progress_bar = CurrentProgressBar(refresh_rate=refresh_rate)
@@ -282,40 +271,40 @@ def test_tqdm_progress_bar_value_on_colab(tmpdir):
     assert trainer.progress_bar_callback.refresh_rate == 19
 
 
-class MockedUpdateProgressBars(TQDMProgressBar):
-    """Mocks the update method once bars get initializied."""
+class MockTqdm(Tqdm):
+    def __init__(self, *args, **kwargs):
+        self.n_values = []
+        super().__init__(*args, **kwargs)
+        self.__n = 0
+        # again to reset additions from `super().__init__`
+        self.n_values = []
 
-    def _mock_bar_update(self, bar):
-        bar.update = Mock(wraps=bar.update)
-        return bar
+    @property
+    def n(self):
+        return self.__n
 
-    def init_train_tqdm(self):
-        bar = super().init_train_tqdm()
-        return self._mock_bar_update(bar)
-
-    def init_validation_tqdm(self):
-        bar = super().init_validation_tqdm()
-        return self._mock_bar_update(bar)
-
-    def init_test_tqdm(self):
-        bar = super().init_test_tqdm()
-        return self._mock_bar_update(bar)
+    @n.setter
+    def n(self, value):
+        self.__n = value
+        # track the changes in the `n` value
+        if not len(self.n_values) or value != self.n_values[-1]:
+            self.n_values.append(value)
 
 
 @pytest.mark.parametrize(
-    "train_batches,val_batches,refresh_rate,train_deltas,val_deltas",
+    "train_batches,val_batches,refresh_rate,train_updates,val_updates",
     [
-        [2, 3, 1, [1, 1, 1, 1, 1], [1, 1, 1]],
-        [0, 0, 3, [], []],
-        [1, 0, 3, [1], []],
+        [2, 3, 1, [1, 2, 3, 4, 5], [1, 2, 3]],
+        [0, 0, 3, None, None],
+        [1, 0, 3, [1], None],
         [1, 1, 3, [2], [1]],
-        [5, 0, 3, [3, 2], []],
-        [5, 2, 3, [3, 3, 1], [2]],
-        [5, 2, 6, [6, 1], [2]],
+        [5, 0, 3, [3, 5], None],
+        [5, 2, 3, [3, 7], [2]],
+        [5, 2, 6, [7], [2]],
     ],
 )
 def test_main_progress_bar_update_amount(
-    tmpdir, train_batches: int, val_batches: int, refresh_rate: int, train_deltas: list, val_deltas: list
+    tmpdir, train_batches: int, val_batches: int, refresh_rate: int, train_updates, val_updates
 ):
     """Test that the main progress updates with the correct amount together with the val progress.
 
@@ -323,7 +312,7 @@ def test_main_progress_bar_update_amount(
     rate.
     """
     model = BoringModel()
-    progress_bar = MockedUpdateProgressBars(refresh_rate=refresh_rate)
+    progress_bar = TQDMProgressBar(refresh_rate=refresh_rate)
     trainer = Trainer(
         default_root_dir=tmpdir,
         max_epochs=1,
@@ -333,18 +322,19 @@ def test_main_progress_bar_update_amount(
         logger=False,
         enable_checkpointing=False,
     )
-    trainer.fit(model)
+    with mock.patch("pytorch_lightning.callbacks.progress.tqdm_progress.Tqdm", MockTqdm):
+        trainer.fit(model)
     if train_batches > 0:
-        progress_bar.main_progress_bar.update.assert_has_calls([call(delta) for delta in train_deltas])
+        assert progress_bar.main_progress_bar.n_values == train_updates
     if val_batches > 0:
-        progress_bar.val_progress_bar.update.assert_has_calls([call(delta) for delta in val_deltas])
+        assert progress_bar.val_progress_bar.n_values == val_updates
 
 
-@pytest.mark.parametrize("test_batches,refresh_rate,test_deltas", [[1, 3, [1]], [3, 1, [1, 1, 1]], [5, 3, [3, 2]]])
-def test_test_progress_bar_update_amount(tmpdir, test_batches: int, refresh_rate: int, test_deltas: list):
+@pytest.mark.parametrize("test_batches,refresh_rate,updates", [[1, 3, [1]], [3, 1, [1, 2, 3]], [5, 3, [3, 5]]])
+def test_test_progress_bar_update_amount(tmpdir, test_batches: int, refresh_rate: int, updates: list):
     """Test that test progress updates with the correct amount."""
     model = BoringModel()
-    progress_bar = MockedUpdateProgressBars(refresh_rate=refresh_rate)
+    progress_bar = TQDMProgressBar(refresh_rate=refresh_rate)
     trainer = Trainer(
         default_root_dir=tmpdir,
         max_epochs=1,
@@ -353,8 +343,9 @@ def test_test_progress_bar_update_amount(tmpdir, test_batches: int, refresh_rate
         logger=False,
         enable_checkpointing=False,
     )
-    trainer.test(model)
-    progress_bar.test_progress_bar.update.assert_has_calls([call(delta) for delta in test_deltas])
+    with mock.patch("pytorch_lightning.callbacks.progress.tqdm_progress.Tqdm", MockTqdm):
+        trainer.test(model)
+    assert progress_bar.test_progress_bar.n_values == updates
 
 
 def test_tensor_to_float_conversion(tmpdir):
@@ -566,36 +557,6 @@ def test_get_progress_bar_metrics(tmpdir: str):
     assert "loss" in standard_metrics.keys()
     assert "split_idx" in standard_metrics.keys()
     assert "v_num" not in standard_metrics.keys()
-
-
-def test_tqdm_progress_bar_main_bar_resume():
-    """Test that the progress bar can resume its counters based on the Trainer state."""
-    bar = TQDMProgressBar()
-    trainer = Mock()
-    model = Mock()
-
-    trainer.sanity_checking = False
-    trainer.check_val_every_n_epoch = 1
-    trainer.current_epoch = 1
-    trainer.num_training_batches = 5
-    trainer.val_check_batch = 5
-    trainer.num_val_batches = [3]
-    trainer.fit_loop.epoch_loop.batch_progress.current.completed = 3
-
-    bar.setup(trainer, model)
-    bar.on_train_start(trainer, model)
-    bar.on_train_epoch_start(trainer, model)
-
-    assert bar.main_progress_bar.n == 3
-    assert bar.main_progress_bar.total == 8
-
-    # bar.on_train_epoch_end(trainer, model)
-    bar.on_validation_start(trainer, model)
-    bar.on_validation_epoch_start(trainer, model)
-
-    # restarting mid validation epoch is not currently supported
-    assert bar.val_progress_bar.n == 0
-    assert bar.val_progress_bar.total == 3
 
 
 def test_tqdm_progress_bar_correct_value_epoch_end(tmpdir):